Initial commit for autofill consolidation task
