--- modules/page_object_autofill_login.py
@@ -1,35 +1,0 @@
-from modules.page_object_autofill import Autofill
-
-
-class LoginAutofill(Autofill):
-    """
-    Page Object Model for the form autofill demo page with many logins
-    """
-
-    URL_TEMPLATE = "https://mozilla.github.io/form-fill-examples/password_manager/login_and_pw_change_forms.html"
-
-    class LoginForm:
-        """
-        Sub class of the Login Autofill Form where you can interact with the Login Form
-        """
-
-        def __init__(self, parent: "LoginAutofill") -> None:
-            self.parent = parent
-            self.username_field = None
-            self.password_field = None
-            self.submit_button = None
-
-        def fill_username(self, username: str) -> None:
-            if self.username_field is None:
-                self.username_field = self.parent.get_element("username-login-field")
-            self.username_field.send_keys(username)
-
-        def fill_password(self, password: str) -> None:
-            if self.password_field is None:
-                self.password_field = self.parent.get_element("password-login-field")
-            self.password_field.send_keys(password)
-
-        def submit(self) -> None:
-            if self.submit_button is None:
-                self.submit_button = self.parent.get_element("submit-button-login")
-            self.submit_button.click()
