--- modules/page_object_form_autofill_textarea.py
@@ -1,9 +1,0 @@
-from modules.page_object_autofill import Autofill
-
-
-class TextAreaFormAutofill(Autofill):
-    """
-    Page Object Model for the form autofill demo page with a textarea
-    """
-
-    URL_TEMPLATE = "https://mozilla.github.io/form-fill-examples/textarea_select.html"
