+++ tests/form_autofill/test_autofill_cc_cvv.py
@@ -6,7 +6,7 @@ from selenium.webdriver import Firefox
 
 from modules.browser_object_autofill_popup import AutofillPopup
 from modules.browser_object_navigation import Navigation
-from modules.page_object_autofill_credit_card import CreditCardFill
+from modules.page_object_autofill import CreditCardFill
 from modules.page_object_prefs import AboutPrefs
 from modules.util import BrowserActions, Utilities
