+++ tests/form_autofill/test_autofill_credit_card_doorhanger.py
@@ -4,7 +4,7 @@ from selenium.webdriver import Firefox
 from modules.browser_object_autofill_popup import AutofillPopup
 from modules.browser_object_navigation import Navigation
 from modules.page_object import AboutPrefs
-from modules.page_object_autofill_credit_card import CreditCardFill
+from modules.page_object_autofill import CreditCardFill
 from modules.util import Utilities
