+++ tests/form_autofill/test_autofill_credit_card_four_fields.py
@@ -3,7 +3,7 @@ from selenium.webdriver import Firefox
 
 from modules.browser_object import Navigation
 from modules.browser_object_autofill_popup import AutofillPopup
-from modules.page_object_autofill_credit_card import CreditCardFill
+from modules.page_object_autofill import CreditCardFill
 from modules.util import Utilities
