+++ tests/form_autofill/test_enable_disable_autofill.py
@@ -3,7 +3,8 @@ from selenium.webdriver import Firefox
 
 from modules.browser_object import Navigation
 from modules.browser_object_autofill_popup import AutofillPopup
-from modules.page_object import AboutPrefs, AddressFill
+from modules.page_object import AboutPrefs
+from modules.page_object_autofill import AddressFill
 from modules.util import Utilities
