+++ tests/form_autofill/test_private_mode_info_not_saved.py
@@ -2,7 +2,8 @@ import pytest
 from selenium.webdriver import Firefox
 
 from modules.browser_object import AutofillPopup
-from modules.page_object import AboutPrefs, AddressFill
+from modules.page_object import AboutPrefs
+from modules.page_object_autofill import AddressFill
 from modules.util import BrowserActions, Utilities
