+++ tests/form_autofill/test_updating_credit_card.py
@@ -7,7 +7,7 @@ from selenium.webdriver import Firefox
 from modules.browser_object import Navigation
 from modules.browser_object_autofill_popup import AutofillPopup
 from modules.page_object import AboutPrefs
-from modules.page_object_autofill_credit_card import CreditCardFill
+from modules.page_object_autofill import CreditCardFill
 from modules.util import BrowserActions, Utilities
