+++ tests/menus/test_copy_paste_actions.py
@@ -6,8 +6,8 @@ from selenium.webdriver.common.keys import Keys
 
 from modules.browser_object import ContextMenu, Navigation
 from modules.page_object import GoogleSearch
-from modules.page_object_autofill_login import LoginAutofill
-from modules.page_object_form_autofill_textarea import TextAreaFormAutofill
+from modules.page_object_autofill import LoginAutofill
+from modules.page_object_autofill import TextAreaFormAutofill
 from modules.util import Utilities
